backend/mcp/requirements.txt
fastapi==0.128.3
mcp==1.26.0
sqlmodel==0.0.16
asyncpg==0.29.0
python-dotenv==1.0.0
uvicorn==0.40.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4